refactor rasp tests
