+++ manifests/cpp.yml
@@ -54,10 +54,8 @@ tests/:
     rasp/:
       test_lfi.py: irrelevant (ASM is not implemented in C++)
       test_shi.py: irrelevant (ASM is not implemented in C++)
-      test_span_tags.py: irrelevant (ASM is not implemented in C++)
       test_sqli.py: irrelevant (ASM is not implemented in C++)
       test_ssrf.py: irrelevant (ASM is not implemented in C++)
-      test_stack_traces.py: irrelevant (ASM is not implemented in C++)
     waf/:
       test_addresses.py: irrelevant (ASM is not implemented in C++)
       test_blocking.py: irrelevant (ASM is not implemented in C++)
