+++ manifests/dotnet.yml
@@ -113,23 +113,27 @@ tests/:
         Test_Lfi_BodyJson: v2.51.0
         Test_Lfi_BodyUrlEncoded: v2.51.0
         Test_Lfi_BodyXml: v2.51.0
+        Test_Lfi_Mandatory_SpanTags: v2.52.0
+        Test_Lfi_Optional_SpanTags: v2.52.0
+        Test_Lfi_StackTrace: v2.51.0
         Test_Lfi_UrlQuery: v2.51.0
       test_shi.py: missing_feature
-      test_span_tags.py:
-        Test_Mandatory_SpanTags: v2.52.0
-        Test_Optional_SpanTags: v2.52.0
       test_sqli.py:
         Test_Sqli_BodyJson: v2.54.0
         Test_Sqli_BodyUrlEncoded: v2.54.0
         Test_Sqli_BodyXml: v2.54.0
+        Test_Sqli_Mandatory_SpanTags: v2.54.0
+        Test_Sqli_Optional_SpanTags: v2.54.0
+        Test_Sqli_StackTrace: v2.54.0
         Test_Sqli_UrlQuery: v2.54.0
       test_ssrf.py:
         Test_Ssrf_BodyJson: v2.51.0
         Test_Ssrf_BodyUrlEncoded: v2.51.0
         Test_Ssrf_BodyXml: v2.51.0
+        Test_Ssrf_Mandatory_SpanTags: v2.51.0
+        Test_Ssrf_Optional_SpanTags: v2.51.0
+        Test_Ssrf_StackTrace: v2.51.0
         Test_Ssrf_UrlQuery: v2.51.0
-      test_stack_traces.py:
-        Test_StackTrace: v2.51.0
     waf/:
       test_addresses.py:
         Test_BodyJson: v2.8.0
