+++ manifests/java.yml
@@ -395,7 +395,6 @@ tests/:
     rasp/:
       test_lfi.py: missing_feature
       test_shi.py: missing_feature
-      test_span_tags.py: missing_feature
       test_sqli.py:
         Test_Sqli_BodyJson:
           '*': v1.38.0
@@ -414,12 +413,14 @@ tests/:
           spring-boot-payara: missing_feature (Requires payara blockng functionallity)
           vertx3: missing_feature (Requires parsed body instrumentation)
           vertx4: missing_feature (Requires parsed body instrumentation)
+        Test_Sqli_Mandatory_SpanTags: missing_feature
+        Test_Sqli_Optional_SpanTags: missing_feature
+        Test_Sqli_StackTrace: missing_feature
         Test_Sqli_UrlQuery:
           '*': v1.38.0
           spring-boot-3-native: missing_feature (GraalVM. Tracing support only)
           spring-boot-payara: missing_feature (Requires payara blockng functionallity)
       test_ssrf.py: missing_feature
-      test_stack_traces.py: missing_feature
     waf/:
       test_addresses.py:
         Test_BodyJson:
