+++ manifests/nodejs.yml
@@ -207,9 +207,6 @@ tests/:
     rasp/:
       test_lfi.py: missing_feature
       test_shi.py: missing_feature
-      test_span_tags.py:
-        Test_Mandatory_SpanTags: *ref_5_18_0
-        Test_Optional_SpanTags: *ref_5_18_0
       test_sqli.py: missing_feature
       test_ssrf.py:
         Test_Ssrf_BodyJson:
@@ -219,11 +216,12 @@ tests/:
           '*': *ref_5_20_0
           nextjs: missing_feature
         Test_Ssrf_BodyXml: missing_feature # xml body not supported
-        Test_Ssrf_UrlQuery:
+        Test_Ssrf_Mandatory_SpanTags: *ref_5_18_0
+        Test_Ssrf_Optional_SpanTags: *ref_5_18_0
+        Test_Ssrf_StackTrace:
           '*': *ref_5_20_0
           nextjs: missing_feature
-      test_stack_traces.py:
-        Test_StackTrace:
+        Test_Ssrf_UrlQuery:
           '*': *ref_5_20_0
           nextjs: missing_feature
     waf/:
