+++ manifests/python.yml
@@ -190,27 +190,34 @@ tests/:
         Test_Lfi_BodyJson: v2.10.0.dev
         Test_Lfi_BodyUrlEncoded: v2.10.0.dev
         Test_Lfi_BodyXml: v2.10.0.dev
+        Test_Lfi_Mandatory_SpanTags: v2.10.0.dev
+        Test_Lfi_Optional_SpanTags: v2.10.0.dev
+        Test_Lfi_StackTrace: v2.10.0.dev
         Test_Lfi_UrlQuery: v2.10.0.dev
       test_shi.py:
         Test_Shi_BodyJson: v2.11.0-rc2
         Test_Shi_BodyUrlEncoded: v2.11.0-rc2
         Test_Shi_BodyXml: v2.11.0-rc2
+        Test_Shi_Mandatory_SpanTags: v2.11.0-rc2
+        Test_Shi_Optional_SpanTags: v2.11.0-rc2
+        Test_Shi_StackTrace: v2.11.0-rc2
         Test_Shi_UrlQuery: v2.11.0-rc2
-      test_span_tags.py:
-        Test_Mandatory_SpanTags: v2.10.0.dev
-        Test_Optional_SpanTags: v2.10.0.dev
       test_sqli.py:
         Test_Sqli_BodyJson: v2.10.0.dev
         Test_Sqli_BodyUrlEncoded: v2.10.0.dev
         Test_Sqli_BodyXml: v2.10.0.dev
+        Test_Sqli_Mandatory_SpanTags: v2.10.0.dev
+        Test_Sqli_Optional_SpanTags: v2.10.0.dev
+        Test_Sqli_StackTrace: v2.10.0.dev
         Test_Sqli_UrlQuery: v2.10.0.dev
       test_ssrf.py:
         Test_Ssrf_BodyJson: v2.10.0.dev
         Test_Ssrf_BodyUrlEncoded: v2.10.0.dev
         Test_Ssrf_BodyXml: v2.10.0.dev
+        Test_Ssrf_Mandatory_SpanTags: v2.10.0.dev
+        Test_Ssrf_Optional_SpanTags: v2.10.0.dev
+        Test_Ssrf_StackTrace: v2.10.0.dev
         Test_Ssrf_UrlQuery: v2.10.0.dev
-      test_stack_traces.py:
-        Test_StackTrace: v2.10.0.dev
     waf/:
       test_addresses.py:
         Test_BodyJson:
