+++ tests/appsec/rasp/test_shi.py
@@ -3,13 +3,14 @@
 # Copyright 2021 Datadog, Inc.
 
 from utils import features, weblog, interfaces, scenarios, rfc
+from tests.appsec.rasp.rasp_utils import validate_span_tags, validate_stack_traces
 
 
 @rfc("https://docs.google.com/document/d/1gCXU3LvTH9en3Bww0AC2coSJWz1m7HcavZjvMLuDCWg/edit#heading=h.giijrtyn1fdx")
 @features.rasp_shell_injection
 @scenarios.appsec_rasp
 class Test_Shi_UrlQuery:
-    """ Shell Injection through query parameters """
+    """Shell Injection through query parameters"""
 
     def setup_shi_get(self):
         self.r = weblog.get("/rasp/shi", params={"list_dir": "$(cat /etc/passwd 1>&2 ; echo .)"})
@@ -21,8 +22,8 @@ class Test_Shi_UrlQuery:
             self.r,
             "rasp-932-100",
             {
-                "resource": {"address": "server.sys.shell.cmd", "value": "ls $(cat /etc/passwd 1>&2 ; echo .)"},
-                "params": {"address": "server.request.query", "value": "$(cat /etc/passwd 1>&2 ; echo .)"},
+                "resource": {"address": "server.sys.shell.cmd", "value": "ls $(cat /etc/passwd 1>&2 ; echo .)",},
+                "params": {"address": "server.request.query", "value": "$(cat /etc/passwd 1>&2 ; echo .)",},
             },
         )
 
@@ -31,7 +32,7 @@ class Test_Shi_UrlQuery:
 @features.rasp_shell_injection
 @scenarios.appsec_rasp
 class Test_Shi_BodyUrlEncoded:
-    """ Shell Injection through a url-encoded body parameter """
+    """Shell Injection through a url-encoded body parameter"""
 
     def setup_shi_post_urlencoded(self):
         self.r = weblog.post("/rasp/shi", data={"list_dir": "$(cat /etc/passwd 1>&2 ; echo .)"})
@@ -43,8 +44,8 @@ class Test_Shi_BodyUrlEncoded:
             self.r,
             "rasp-932-100",
             {
-                "resource": {"address": "server.sys.shell.cmd", "value": "ls $(cat /etc/passwd 1>&2 ; echo .)"},
-                "params": {"address": "server.request.body", "value": "$(cat /etc/passwd 1>&2 ; echo .)"},
+                "resource": {"address": "server.sys.shell.cmd", "value": "ls $(cat /etc/passwd 1>&2 ; echo .)",},
+                "params": {"address": "server.request.body", "value": "$(cat /etc/passwd 1>&2 ; echo .)",},
             },
         )
 
@@ -53,7 +54,7 @@ class Test_Shi_BodyUrlEncoded:
 @features.rasp_shell_injection
 @scenarios.appsec_rasp
 class Test_Shi_BodyXml:
-    """ Shell Injection through an xml body parameter """
+    """Shell Injection through an xml body parameter"""
 
     def setup_shi_post_xml(self):
         data = "<?xml version='1.0' encoding='utf-8'?><list_dir>$(cat /etc/passwd 1>&amp;2 ; echo .)</list_dir>"
@@ -66,8 +67,8 @@ class Test_Shi_BodyXml:
             self.r,
             "rasp-932-100",
             {
-                "resource": {"address": "server.sys.shell.cmd", "value": "ls $(cat /etc/passwd 1>&2 ; echo .)"},
-                "params": {"address": "server.request.body", "value": "$(cat /etc/passwd 1>&2 ; echo .)"},
+                "resource": {"address": "server.sys.shell.cmd", "value": "ls $(cat /etc/passwd 1>&2 ; echo .)",},
+                "params": {"address": "server.request.body", "value": "$(cat /etc/passwd 1>&2 ; echo .)",},
             },
         )
 
@@ -76,7 +77,7 @@ class Test_Shi_BodyXml:
 @features.rasp_shell_injection
 @scenarios.appsec_rasp
 class Test_Shi_BodyJson:
-    """ Shell Injection through a json body parameter """
+    """Shell Injection through a json body parameter"""
 
     def setup_shi_post_json(self):
         """AppSec detects attacks in JSON body values"""
@@ -89,7 +90,52 @@ class Test_Shi_BodyJson:
             self.r,
             "rasp-932-100",
             {
-                "resource": {"address": "server.sys.shell.cmd", "value": "ls $(cat /etc/passwd 1>&2 ; echo .)"},
-                "params": {"address": "server.request.body", "value": "$(cat /etc/passwd 1>&2 ; echo .)"},
+                "resource": {"address": "server.sys.shell.cmd", "value": "ls $(cat /etc/passwd 1>&2 ; echo .)",},
+                "params": {"address": "server.request.body", "value": "$(cat /etc/passwd 1>&2 ; echo .)",},
             },
         )
+
+
+@rfc("https://docs.google.com/document/d/1vmMqpl8STDk7rJnd3YBsa6O9hCls_XHHdsodD61zr_4/edit#heading=h.96mezjnqf46y")
+@features.rasp_span_tags
+@features.rasp_shell_injection
+@scenarios.appsec_rasp
+class Test_Shi_Mandatory_SpanTags:
+    """Validate span tag generation on exploit attempts"""
+
+    def setup_shi_span_tags(self):
+        self.r = weblog.get("/rasp/shi", params={"list_dir": "$(cat /etc/passwd 1>&2 ; echo .)"})
+
+    def test_shi_span_tags(self):
+        validate_span_tags(self.r, expected_metrics=["_dd.appsec.rasp.duration"])
+
+
+@rfc("https://docs.google.com/document/d/1vmMqpl8STDk7rJnd3YBsa6O9hCls_XHHdsodD61zr_4/edit#heading=h.96mezjnqf46y")
+@features.rasp_span_tags
+@features.rasp_shell_injection
+@scenarios.appsec_rasp
+class Test_Shi_Optional_SpanTags:
+    """Validate span tag generation on exploit attempts"""
+
+    def setup_shi_span_tags(self):
+        self.r = weblog.get("/rasp/shi", params={"list_dir": "$(cat /etc/passwd 1>&2 ; echo .)"})
+
+    def test_shi_span_tags(self):
+        validate_span_tags(
+            self.r, expected_metrics=["_dd.appsec.rasp.duration_ext", "_dd.appsec.rasp.rule.eval",],
+        )
+
+
+@rfc("https://docs.google.com/document/d/1vmMqpl8STDk7rJnd3YBsa6O9hCls_XHHdsodD61zr_4/edit#heading=h.enmf90juqidf")
+@features.rasp_stack_trace
+@features.rasp_shell_injection
+@scenarios.appsec_rasp
+class Test_Shi_StackTrace:
+    """Validate stack trace generation on exploit attempts"""
+
+    def setup_shi_stack_trace(self):
+        self.r = weblog.get("/rasp/shi", params={"list_dir": "$(cat /etc/passwd 1>&2 ; echo .)"})
+
+    def test_shi_stack_trace(self):
+        assert self.r.status_code == 403
+        validate_stack_traces(self.r)
