+++ tests/appsec/rasp/test_ssrf.py
@@ -3,13 +3,14 @@
 # Copyright 2021 Datadog, Inc.
 
 from utils import features, weblog, interfaces, scenarios, rfc, context
+from tests.appsec.rasp.rasp_utils import validate_span_tags, validate_stack_traces
 
 
 @rfc("https://docs.google.com/document/d/1vmMqpl8STDk7rJnd3YBsa6O9hCls_XHHdsodD61zr_4/edit#heading=h.3r1lwuv4y2g3")
 @features.rasp_server_side_request_forgery
 @scenarios.appsec_rasp
 class Test_Ssrf_UrlQuery:
-    """ Server-side request forgery through query parameters """
+    """Server-side request forgery through query parameters"""
 
     def setup_ssrf_get(self):
         self.r = weblog.get("/rasp/ssrf", params={"domain": "169.254.169.254"})
@@ -25,8 +26,8 @@ class Test_Ssrf_UrlQuery:
             self.r,
             "rasp-934-100",
             {
-                "resource": {"address": "server.io.net.url", "value": expected_http_value},
-                "params": {"address": "server.request.query", "value": "169.254.169.254"},
+                "resource": {"address": "server.io.net.url", "value": expected_http_value,},
+                "params": {"address": "server.request.query", "value": "169.254.169.254",},
             },
         )
 
@@ -35,7 +36,7 @@ class Test_Ssrf_UrlQuery:
 @features.rasp_server_side_request_forgery
 @scenarios.appsec_rasp
 class Test_Ssrf_BodyUrlEncoded:
-    """ Server-side request forgery through a url-encoded body parameter """
+    """Server-side request forgery through a url-encoded body parameter"""
 
     def setup_ssrf_post_urlencoded(self):
         self.r = weblog.post("/rasp/ssrf", data={"domain": "169.254.169.254"})
@@ -51,8 +52,8 @@ class Test_Ssrf_BodyUrlEncoded:
             self.r,
             "rasp-934-100",
             {
-                "resource": {"address": "server.io.net.url", "value": expected_http_value},
-                "params": {"address": "server.request.body", "value": "169.254.169.254"},
+                "resource": {"address": "server.io.net.url", "value": expected_http_value,},
+                "params": {"address": "server.request.body", "value": "169.254.169.254",},
             },
         )
 
@@ -61,10 +62,10 @@ class Test_Ssrf_BodyUrlEncoded:
 @features.rasp_server_side_request_forgery
 @scenarios.appsec_rasp
 class Test_Ssrf_BodyXml:
-    """ Server-side request forgery through an xml body parameter """
+    """Server-side request forgery through an xml body parameter"""
 
     def setup_ssrf_post_xml(self):
-        data = f"<?xml version='1.0' encoding='utf-8'?><domain>169.254.169.254</domain>"
+        data = "<?xml version='1.0' encoding='utf-8'?><domain>169.254.169.254</domain>"
         self.r = weblog.post("/rasp/ssrf", data=data, headers={"Content-Type": "application/xml"})
 
     def test_ssrf_post_xml(self):
@@ -74,8 +75,8 @@ class Test_Ssrf_BodyXml:
             self.r,
             "rasp-934-100",
             {
-                "resource": {"address": "server.io.net.url", "value": "http://169.254.169.254"},
-                "params": {"address": "server.request.body", "value": "169.254.169.254"},
+                "resource": {"address": "server.io.net.url", "value": "http://169.254.169.254",},
+                "params": {"address": "server.request.body", "value": "169.254.169.254",},
             },
         )
 
@@ -84,7 +85,7 @@ class Test_Ssrf_BodyXml:
 @features.rasp_server_side_request_forgery
 @scenarios.appsec_rasp
 class Test_Ssrf_BodyJson:
-    """ Server-side request forgery through a json body parameter """
+    """Server-side request forgery through a json body parameter"""
 
     def setup_ssrf_post_json(self):
         """AppSec detects attacks in JSON body values"""
@@ -101,7 +102,52 @@ class Test_Ssrf_BodyJson:
             self.r,
             "rasp-934-100",
             {
-                "resource": {"address": "server.io.net.url", "value": expected_http_value},
-                "params": {"address": "server.request.body", "value": "169.254.169.254"},
+                "resource": {"address": "server.io.net.url", "value": expected_http_value,},
+                "params": {"address": "server.request.body", "value": "169.254.169.254",},
             },
         )
+
+
+@rfc("https://docs.google.com/document/d/1vmMqpl8STDk7rJnd3YBsa6O9hCls_XHHdsodD61zr_4/edit#heading=h.96mezjnqf46y")
+@features.rasp_span_tags
+@features.rasp_server_side_request_forgery
+@scenarios.appsec_rasp
+class Test_Ssrf_Mandatory_SpanTags:
+    """Validate span tag generation on exploit attempts"""
+
+    def setup_ssrf_span_tags(self):
+        self.r = weblog.get("/rasp/ssrf", params={"domain": "169.254.169.254"})
+
+    def test_ssrf_span_tags(self):
+        validate_span_tags(self.r, expected_metrics=["_dd.appsec.rasp.duration"])
+
+
+@rfc("https://docs.google.com/document/d/1vmMqpl8STDk7rJnd3YBsa6O9hCls_XHHdsodD61zr_4/edit#heading=h.96mezjnqf46y")
+@features.rasp_span_tags
+@features.rasp_server_side_request_forgery
+@scenarios.appsec_rasp
+class Test_Ssrf_Optional_SpanTags:
+    """Validate span tag generation on exploit attempts"""
+
+    def setup_ssrf_span_tags(self):
+        self.r = weblog.get("/rasp/ssrf", params={"domain": "169.254.169.254"})
+
+    def test_ssrf_span_tags(self):
+        validate_span_tags(
+            self.r, expected_metrics=["_dd.appsec.rasp.duration_ext", "_dd.appsec.rasp.rule.eval",],
+        )
+
+
+@rfc("https://docs.google.com/document/d/1vmMqpl8STDk7rJnd3YBsa6O9hCls_XHHdsodD61zr_4/edit#heading=h.enmf90juqidf")
+@features.rasp_stack_trace
+@features.rasp_server_side_request_forgery
+@scenarios.appsec_rasp
+class Test_Ssrf_StackTrace:
+    """Validate stack trace generation on exploit attempts"""
+
+    def setup_ssrf_stack_trace(self):
+        self.r = weblog.get("/rasp/ssrf", params={"domain": "169.254.169.254"})
+
+    def test_ssrf_stack_trace(self):
+        assert self.r.status_code == 403
+        validate_stack_traces(self.r)
